feat: 다양한 타입의 아이템 만들기
- 아이템 별로 외관이 다르다.
- 이동속도: 플레이어의 이동속도 up
- 체력: 플레이어의 체력 up
- 공격속도 : 플레이어의 공격속도 u

+++ Assets/02.Scripts/Player/PlayerMove.cs
@@ -34,7 +34,7 @@ public class PlayerMove : MonoBehaviour
         _originPosition = transform.position;
     }
 
-    public void SpeedUp(int value)
+    public void SpeedUp(float value)
     {
         _speed += value;
